Refactor remotes into submodule.
